+++ meshers/blocky/voxel_blocky_library.h
@@ -2,7 +2,7 @@
 #define VOXEL_LIBRARY_H
 
 #include "../../util/dynamic_bitset.h"
-#include "voxel.h"
+#include "voxel_blocky_model.h"
 #include <core/object/ref_counted.h>
 
 namespace zylann::voxel {
@@ -10,8 +10,8 @@ namespace zylann::voxel {
 // TODO Rename VoxelBlockyLibrary
 
 // Stores a list of models that can be used with VoxelMesherBlocky
-class VoxelLibrary : public Resource {
-	GDCLASS(VoxelLibrary, Resource)
+class VoxelBlockyLibrary : public Resource {
+	GDCLASS(VoxelBlockyLibrary, Resource)
 
 public:
 	// Limit based on maximum supported by VoxelMesherBlocky
@@ -24,7 +24,7 @@ public:
 		DynamicBitset side_pattern_culling;
 		unsigned int side_pattern_count = 0;
 		// Lots of data can get moved but it's only on load.
-		std::vector<Voxel::BakedData> models;
+		std::vector<VoxelBlockyModel::BakedData> models;
 
 		inline bool has_model(uint32_t i) const {
 			return i < models.size();
@@ -39,8 +39,8 @@ public:
 		}
 	};
 
-	VoxelLibrary();
-	~VoxelLibrary();
+	VoxelBlockyLibrary();
+	~VoxelBlockyLibrary();
 
 	int get_atlas_size() const {
 		return _atlas_size;
@@ -48,7 +48,7 @@ public:
 	void set_atlas_size(int s);
 
 	// Use this factory rather than creating voxels from scratch
-	Ref<Voxel> create_voxel(unsigned int id, String name);
+	Ref<VoxelBlockyModel> create_voxel(unsigned int id, String name);
 
 	unsigned int get_voxel_count() const;
 	void set_voxel_count(unsigned int type_count);
@@ -71,7 +71,7 @@ public:
 		return id < _voxel_types.size() && _voxel_types[id].is_valid();
 	}
 
-	_FORCE_INLINE_ const Voxel &get_voxel_const(unsigned int id) const {
+	_FORCE_INLINE_ const VoxelBlockyModel &get_voxel_const(unsigned int id) const {
 		return **_voxel_types[id];
 	}
 
@@ -83,7 +83,7 @@ public:
 	}
 
 private:
-	void set_voxel(unsigned int id, Ref<Voxel> voxel);
+	void set_voxel(unsigned int id, Ref<VoxelBlockyModel> voxel);
 
 	void generate_side_culling_matrix();
 
@@ -91,15 +91,15 @@ private:
 	bool _get(const StringName &p_name, Variant &r_ret) const;
 	void _get_property_list(List<PropertyInfo> *p_list) const;
 
-	Ref<Voxel> _b_get_voxel(unsigned int id);
-	Ref<Voxel> _b_get_voxel_by_name(StringName name);
+	Ref<VoxelBlockyModel> _b_get_voxel(unsigned int id);
+	Ref<VoxelBlockyModel> _b_get_voxel_by_name(StringName name);
 
 	static void _bind_methods();
 
 private:
 	// There can be null entries. A vector is used because there should be no more than 65,536 items,
 	// and in practice the intented use case rarely goes over a few hundreds
-	std::vector<Ref<Voxel>> _voxel_types;
+	std::vector<Ref<VoxelBlockyModel>> _voxel_types;
 	int _atlas_size = 16;
 	bool _needs_baking = true;
 	bool _bake_tangents = true;